requirements/ci.txt
pyfirecrest==3.0.0